Homepage complete and courses progress
I got a lot of work done on the homepage a courses page. The homepage is complete, but the courses page still needs quite a bit of work. The login page has not been worked on at all. Also, the courses.js in the data folder still needs to be filled in.